QBAP Improvement, Boost Fix, Campaign Selection Emojis

+++ cogs/misc.py
@@ -15,6 +15,8 @@ class Misc(commands.Cog, name='Misc'):
         self.quest_board_channel_id = 382027190633627649 #382027190633627649 725577624180621313
         self.category_channel_id = 382027737189056544 #382027737189056544  728456686024523810
 
+
+    #https://discordapp.com/channels/382025597041246210/432358370578530310/733403065251528795 nice comments that make it worth it <3
     @commands.cooldown(1, 60, type=commands.BucketType.member)
     @commands.command()
     async def uwu(self,ctx):
@@ -52,7 +54,6 @@ class Misc(commands.Cog, name='Misc'):
         await channel.send(content=message.author.display_name + ":\n" +  uwuMessage)
         await ctx.message.delete()
         
-    
 
     #searches for the last message sent by the bot in case a restart was made
     #Allows it to use it to remove the last post
@@ -183,9 +184,10 @@ class Misc(commands.Cog, name='Misc'):
         game_channel_ids = set(map(lambda c: c.id, game_channel_category.text_channels))
         build_message = "The current status of the game channels is:\n"
         #create a dictonary to store the room/user pairs
+        tierMap = {"Tier 0" : "T0", "Tier 1" : "T1", "Tier 2" : "T2", "Tier 3" : "T3", "Tier 4" : "T4"}
         channel_dm_dic = {}
         for c in game_channel_category.text_channels:
-            channel_dm_dic[c.mention]= 'Clear'
+            channel_dm_dic[c.mention]= ["✅ "+c.mention+": Clear", set([])]
         #get all posts in the channel
         all_posts = await channel.history(oldest_first=True).flatten()
         for elem in all_posts:
@@ -195,10 +197,23 @@ class Misc(commands.Cog, name='Misc'):
             #loop in order to avoid guild channels blocking the check
             for mention in elem.channel_mentions:
                 if mention.id in game_channel_ids:
-                    channel_dm_dic[mention.mention] = elem.author.name
+                    username = elem.author.name
+                    if(elem.author.nick):
+                        username = elem.author.nick
+                    channel_dm_dic[mention.mention][0] = "❌ "+mention.mention+": "+username
+                    for tierMention in elem.role_mentions:
+                        print(tierMention)
+                        if tierMention.name in tierMap:
+                            channel_dm_dic[mention.mention][1].add(tierMap[tierMention.name])
         #build the message using the pairs built above
+        print(channel.guild.me)
         for c in game_channel_category.text_channels:
-            build_message+=c.mention+": "+channel_dm_dic[c.mention]+"\n"
+            print(c, c.permissions_for(channel.guild.me).view_channel)
+            if(c.permissions_for(channel.guild.me).view_channel):
+                tierAddendum = ""
+                if(len(channel_dm_dic[c.mention][1])> 0):
+                    tierAddendum = " - "+"/".join(sorted(channel_dm_dic[c.mention][1]))
+                build_message+=channel_dm_dic[c.mention][0]+tierAddendum+"\n"
         return build_message
     
         
@@ -230,7 +245,7 @@ class Misc(commands.Cog, name='Misc'):
             if(self.current_message and self.past_message_check != 1):
                 #in case a message is posted without a game channel which is then edited in we need to this extra check
                 msgAfter = False
-                async for message in ctx.channel.history(after=self.current_message, limit=1):
+                async for message in self.bot.get_channel(self.quest_board_channel_id).history(after=self.current_message, limit=1):
                     msgAfter = True
                 if( not msgAfter):
                     await self.current_message.edit(content=new_text)
@@ -238,10 +253,18 @@ class Misc(commands.Cog, name='Misc'):
                     await self.current_message.delete()
                     self.current_message = await self.current_message.send(content=new_text)
             else:
-                if(self.current_message):
-                    await self.current_message.delete()
                 self.past_message_check = 2
+                if(self.current_message):                
+                    msgAfter = False
+                    async for message in self.bot.get_channel(self.quest_board_channel_id).history(after=self.current_message, limit=1):
+                        msgAfter = True
+                    if(not msgAfter):
+                        await self.current_message.edit(content=new_text)
+                        return
+                    else:
+                        await self.current_message.delete()
                 self.current_message = await self.bot.get_channel(self.quest_board_channel_id).send(content=new_text)
+                
             
     @commands.Cog.listener()
     async def on_message(self,msg):
@@ -282,7 +305,7 @@ class Misc(commands.Cog, name='Misc'):
             await msg.add_reaction('👋')
         #check if any tier boost was done and react
         elif(7 < msg.type.value and msg.type.value < 12):
-            await msg.add_reaction(':boost:')
+            await msg.add_reaction('<:boost:585637770970660876>')
         elif any(word in msg.content.lower() for word in ['thank', 'thanks', 'thank you', 'thx', 'gracias', 'danke']) and 'bot friend' in msg.content.lower():
             await msg.add_reaction('❤️')
             await msg.channel.send("You're welcome friend!")
@@ -295,20 +318,14 @@ class Misc(commands.Cog, name='Misc'):
             game_channel_ids = list(map(lambda c: c.id, game_channel_category.text_channels))
             for mention in cMentionArray:
                 if mention.id in game_channel_ids:
+                    new_text = await (self.generateMessageText)()
                     if(self.past_message_check == 2):
-                        #since we have the latest information and know who asked for the channel we can just update that line instead of searching through posts
-                        def generateMessageText(currentMessage, target_channel):
-                            reg = target_channel.mention+".*?\n"
-                            repl = target_channel.mention+": "+msg.author.name+"\n"
-                            build_message = re.sub(reg,repl, currentMessage.content)
-                            return build_message
                         await self.current_message.delete()
-                        self.current_message = await msg.channel.send(content=generateMessageText(self.current_message, mention))
+                        self.current_message = await msg.channel.send(content=new_text)
                         return
                     #if there is an old message our record could be out of date so we need to regather info and go to the bottom
                     elif(self.past_message_check == 1 and self.current_message):
                         await self.current_message.delete()
-                    new_text = await (self.generateMessageText)()
                     self.past_message_check = 2
                     self.current_message = await msg.channel.send(content=new_text)
                     return
